requirements.txt
cffi==1.1.2
cryptography==0.9.1
docker-py==1.2.2
idna==2.0
pyasn1==0.1.7
pycparser==2.14
pyOpenSSL==0.15.1
python-etcd==0.3.3
requests==2.7.0
six==1.9.0
urllib3==1.10.4
